Split up compiler and helpers and added a get timepoint function
